accessibility: Transport accessibility measures

A set of fast and convenient functions to calculate multiple transport accessibility measures. Given a pre-computed travel cost matrix and a land use dataset (containing the location of jobs, healthcare and population, for example), the package allows one to calculate active and passive accessibility levels using multiple accessibility measures, such as: cumulative opportunities (using either travel cost cutoffs or intervals), minimum travel cost to closest N number of activities, gravity-based (with different decay functions) and different floating catchment area methods.

Usage

Please check the vignettes for more on the package usage:

Introduction to accessibility: calculating accessibility measures. Run vignette("accessibility") or check it on the website.

Decay functions. Run vignette("decay_functions", package = "accessibility") or check it on the website.

Calculating accessibility inequality and poverty. Run vignette("inequality_and_poverty", package = "accessibility") or check it on the website.

See also

Author

Maintainer: Daniel Herszenhut dhersz@gmail.com (ORCID)

Authors:

Rafael H. M. Pereira rafa.pereira.br@gmail.com (ORCID)

Other contributors:

Ipea - Institute for Applied Economic Research [copyright holder, funder]