Concentration index

Calculates the Concentration Index to measure the socioeconomic inequality of access to opportunities. It measures the extent to which inequalities in accessibility are systematically associated with individuals' socioeconomic levels. The function currently include the options of calculating the corrected concentration index proposed by Erreygers (2009) and the standard relative concentration index.

concentration_index(
  accessibility_data,
  sociodemographic_data,
  opportunity,
  population,
  income,
  type = "CIc",
  group_by = character(0)
)

Arguments

accessibility_data

A data frame. The accessibility levels whose inequality should be calculated. Must contain the columns id and any others specified in opportunity.

sociodemographic_data

A data frame. The distribution of sociodemographic characteristics of the population in the study area cells. Must contain the columns id and any others specified in population and income.

opportunity

A string. The name of the column in accessibility_data with the accessibility levels to be considerend when calculating inequality levels.

population

A string. The name of the column in sociodemographic_data with the number of people in each cell. Used to weigh accessibility levels when calculating inequality.

income

A string. The name of column in sociodemographic_data with the income variable that should be used to classify the population in socioeconomic groups. Please note that this variable should describe income per capita (e.g. mean income per capita, household income per capita, etc), instead of the total amount of income in each cell. ?? income or SES ? daniel

type

A string. the type of concentration index to be calculated. Options include "CIc" for the Corrected concentration index proposed by Erreygers (2009) (the default), and "CI" for the the standard relative concentration index.

group_by

A character vector. When not character(0) (the default), indicates the accessibility_data columns that should be used to group the poverty estimates by. For example, if accessibility_data includes a "scenario" column that specifies the accessibility estimates "before" and "after" given a transport policy intervention, passing "scenario" to this parameter results in inequality estimates for each scenario.

Value

A data frame containing the inequality estimates for the study area.

Interpretation of the concentration index

The concentration index can theoretically vary between -1 and +1 (when all accessibility is concentrated in the most or in the least disadvantaged person, respectively). Negative values indicate that inequalities favor the poor, while positive values indicate a pro-rich bias.

References

Erreygers G (2009). “Correcting the Concentration Index.” Journal of Health Economics, 28(2), 504--515. ISSN 0167-6296, doi:10.1016/j.jhealeco.2008.02.003 , 2023-06-25.

See also

Examples

data_dir <- system.file("extdata", package = "accessibility")
travel_matrix <- readRDS(file.path(data_dir, "travel_matrix.rds"))
land_use_data <- readRDS(file.path(data_dir, "land_use_data.rds"))

access <- cumulative_cutoff(
  travel_matrix,
  land_use_data,
  cutoff = 30,
  opportunity = "jobs",
  travel_cost = "travel_time"
)

ci <- concentration_index(
  accessibility_data = access,
  sociodemographic_data = land_use_data,
  opportunity = "jobs",
  population = "population",
  income = "income_per_capita",
  type = "CIc"
)
ci
#>           CI
#> 1: 0.3346476