Cumulative access based on a travel cost cutoff

Calculates the number of opportunities accessible under a given specified travel cost cutoff.

This function is generic over any kind of numeric travel cost, such as distance, time and money.

cumulative_time_cutoff(
  travel_matrix,
  land_use_data,
  cutoff,
  opportunity_col,
  travel_cost_col = "travel_time",
  by_col = NULL,
  active = TRUE
)

Arguments

travel_matrix

A data frame. The travel matrix describing the costs (i.e. travel time, distance, monetary cost, etc.) between the origins and destinations in the study area. Must contain the columns from_id, to_id and any others specified in travel_cost_col.

land_use_data

A data frame. The distribution of opportunities within the study area cells. Must contain the columns id and any others specified in opportunity_col.

cutoff

A numeric. A number indicating the travel cost cutoff.

opportunity_col

A string. The name of the column in land_use_data with the number of opportunities/resources/services to be considered when calculating accessibility levels.

travel_cost_col

A string. The name of the column in travel_matrix with the travel cost between origins and destinations. Defaults to "travel_time".

by_col

Either NULL (the default) or a character vector. When not NULL, indicates the travel_matrix columns that should be used to group the accessibility estimates by. For example, if travel_matrix includes a departure time column, that specifies the departure time of each entry in the data frame, passing "departure_time" to this argument results in accessibility estimates grouped by origin and by departure time.

active

A logical. Whether to calculate active accessibility (the of opportunities that can be reached from a given origin, the default) or passive accessibility (how many opportunities/people can reach a given destination).

Value

A data.table containing the accessibility estimate for each origin in the travel matrix.

See also

Examples

data_dir <- system.file("extdata", package = "accessibility")
travel_matrix <- readRDS(file.path(data_dir, "travel_matrix.rds"))
land_use_data <- readRDS(file.path(data_dir, "land_use_data.rds"))

# active accessibility: number of schools accessible from each origin
df <- cumulative_time_cutoff(
  travel_matrix = travel_matrix,
  land_use_data = land_use_data,
  cutoff = 30,
  opportunity_col = "schools",
  travel_cost_col = "travel_time"
)
head(df)
#>                 id schools
#> 1: 89a88cdb57bffff       6
#> 2: 89a88cdb597ffff      22
#> 3: 89a88cdb5b3ffff      23
#> 4: 89a88cdb5cfffff       6
#> 5: 89a88cd909bffff      17
#> 6: 89a88cd90b7ffff      20

# passive accessibility: number of people that can reach each destination
df <- cumulative_time_cutoff(
  travel_matrix = travel_matrix,
  land_use_data = land_use_data,
  cutoff = 30,
  opportunity_col = "population",
  travel_cost_col = "travel_time",
  active = FALSE
)
head(df)
#>                 id population
#> 1: 89a88cdb57bffff      62591
#> 2: 89a88cdb19bffff     180316
#> 3: 89a88cdb1d3ffff     185293
#> 4: 89a88cdb56fffff     162202
#> 5: 89a88cdb0afffff     243124
#> 6: 89a88cdb50fffff     107605