Cumulative access based on maximum travel time interval

The function calculates the average or median number of opportunities that can be reached considering multiple minute-by-minute maximum travel time thresholds within a given travel time interval specified by the user.

cumulative_time_interval(
  data,
  opportunity_col,
  interval,
  by_col,
  stat = "mean"
)

Arguments

data

A data.frame with a travel time matrix in long format, containing the at least the columns of origin from_id, destination to_id, travel time travel_time from origin to destination, and number of opportunities in destination locations.

opportunity_col

A string indicating the name of the column with data on the opportunities to be considered.

interval

An numeric vector of length 2, indicating the start and end points of the interval of travel time thresholds to be used.

by_col

A string with the name of the column of origin or destination that should be considered, indicating whether accessibility levels should by calculated by each origin (active accessibility) or destination (passive accessibility).

stat

A string indicating the summary statistic used to aggregate the accessibility estimates within the time interval. It accepts either median (Default) or mean.

Value

A data.table object.

See also

Examples

library(accessibility)

# load a travel time matrix data in long format
data_path <- system.file("extdata/ttm_bho.rds", package = "accessibility")
ttm <- readRDS(data_path)

df <- cumulative_time_interval(data = ttm,
                              opportunity_col = 'schools',
                              interval = c(20, 30),
                              by_col = 'from_id',
                              stat ='mean')
head(df)
#>            from_id     access
#> 1: 89a88cdb57bffff  0.8181818
#> 2: 89a88cdb597ffff 15.0000000
#> 3: 89a88cdb5b3ffff 17.1818182
#> 4: 89a88cdb5cfffff  3.8181818
#> 5: 89a88cd909bffff  8.5454545
#> 6: 89a88cd90b7ffff 12.7272727