Foster–Greer–Thorbecke (FGT) poverty measures

Calculates the Foster–Greer–Thorbecke (FGT) indices, a family of poverty measures that capture the extent and severity of poverty. The FGT family comprises three indicators that differ based on the the alpha parameter, which can receive values of 0, 1 or 2. The interpretation of FGT is different for each value of alpha. With alpha = 0, FGT0 captures the extent of poverty as a simple headcount, i.e. the proportion of people below the poverty line. With alpha = 1, FGT1 captures the severity of poverty as the average percent distance between the poverty line and the accessibility of individuals below the poverty line. FGT1 is also known as the “poverty gap index”. Finally, with alpha = 2, FGT2 index simultaneously captures the extent and severity of poverty by calculating the number of people below the poverty line weighted by the size of the accessibility shortfall relative to the poverty line. The value of FGT2 ranges from 0 to 1, where 0 indicates that every individual is above the poverty line, and a value of 1 indicates that the entire population is below the poverty line. The function returns a data.frame with the values of FGT0, FGT1 and FGT2. The FGT family of poverty measures was originally proposed by Foster et al. (1984) .

fgt_poverty(
  accessibility_data,
  opportunity,
  sociodemographic_data = NULL,
  population = NULL,
  poverty_line,
  group_by = character(0)
)

Arguments

accessibility_data

A data frame with accessibility estimates. It must contain the columns id and any others specified in opportunity.

opportunity

A string. The name of the column in accessibility_data with the accessibility estimates to be considered when calculating accessibility poverty.

sociodemographic_data

A data frame. The distribution of the population in the study area cells. Must contain the columns id and any others specified in population.

population

A string. The name of the column in sociodemographic_data with the number of people in each cell. Used to weigh accessibility levels when calculating poverty.

poverty_line

A numeric value. The value of the accessibility poverty line, below which individuals are considered to be in accessibility poverty.

group_by

A character vector. When not character(0) (the default), indicates the accessibility_data columns that should be used to group the poverty estimates by. For example, if accessibility_data includes a race column that specifies the racial category of the population (e.g. "Black" and "White") that each entry refers to, passing "race" to this parameter results in poverty estimates grouped by race.

Value

A data frame containing the poverty three estimates for the study area: FGT0, FGT1 and FGT2.

Interpretation of Alpha values:

With alpha = 0, the FGT0 index captures the extent of poverty as a simple headcount, i.e. the proportion of people below the poverty line.

With alpha = 1, FGT1 captures the severity of poverty as the average percent distance between the poverty line and the accessibility of individuals below the poverty line. It is also known as the “poverty gap index”.

Finally, with alpha = 2, the FGT2 index simultaneously captures the extent and severity of poverty by calculating the number of people below the poverty line weighted by the size of the accessibility shortfall relative to the poverty line. The value of FGT2 ranges from 0 to 1, where 0 indicates that every individual is above the poverty line, and a value of 1 indicates that the entire population is below the poverty line.

Examples

data_dir <- system.file("extdata", package = "accessibility")
travel_matrix <- readRDS(file.path(data_dir, "travel_matrix.rds"))
land_use_data <- readRDS(file.path(data_dir, "land_use_data.rds"))

access <- cumulative_cutoff(
  travel_matrix = travel_matrix,
  land_use_data = land_use_data,
  cutoff = 30,
  opportunity = "jobs",
  travel_cost = "travel_time"
)

poverty <- fgt_poverty(
  accessibility_data = access,
  opportunity = "jobs",
  sociodemographic_data = land_use_data,
  population = "population",
  poverty_line = 95368
)
poverty
#>         FGT0      FGT1      FGT2
#> 1: 0.5745378 0.3277383 0.2218769