Floating catchment area accessibility

Calculates accessibility accounting for the competition of resources using one of the multiple accessibility metrics in the floating catchment area family. The function currently includes 2SFCA 2-Step Floating Catchment Area and BFCA Balanced Floating Catchment Area. Accessibility can be calculated using multiple alternative decay functions passed through the decay_function parameter.

floating_catchment_area(
  data,
  fca_metric,
  orig_col,
  dest_col,
  population_col,
  opportunity_col,
  decay_function,
  travel_cost_col = "travel_time"
)

Arguments

data

A data.frame with a travel cost matrix in long format, containing the columns with origin ids, destination ids, the travel cost between each origin-destination pair (this can be either in terms of travel time, distance or monetary costs), and number of population and opportunities located in each origin and destination.

fca_metric

A string indicating which floating catchment area accessibility measure to use. Options available: "2SFCA" and "BFCA". More info in the details.

orig_col

A string with the name of the column of origin ids.

dest_col

A string with the name of the column of destination ids.

population_col

A string with the name of the column of origin with population count.

opportunity_col

A string indicating the name of the column with the number of opportunities / resources / services to be considered.

decay_function

A fuction that converts travel cost (t_ij) into an impedance factor used to weight opportunities. For convenience, the package currently includes the following functions: decay_bineary, decay_exponential, decay_exponential and decay_power. See the documentation of each decay function for more details.

travel_cost_col

A string with the name of the column with travel cost between origins and destinations. It can be either either travel time, distance or monetary costs. Defaults to travel_ime.

Value

A data.table object.

2SFCA

The 2SFCA measure was the first accessibility metric in the floating catchment area family. It was originally proposed by Luo & Wang (2003).

BFCA

The balanced floating catchment area (BFCA) metric calculates accessibility accounting for competition effects while simultaneously correcting for issues of inflation of demand and service levels that are present in previous floating catchment area measures. BFCA was was originally proposed by Paez et al. (2019) and named in Pereira et al. (2021).

References:

Luo, W., & Wang, F. (2003). Measures of spatial accessibility to health care in a GIS environment: synthesis and a case study in the Chicago region. Environment and planning B: planning and design, 30(6), 865-884. doi:10.1068/b29120 .

Paez, A., Higgins, C. D., & Vivona, S. F. (2019). Demand and level of service inflation in Floating Catchment Area (FCA) methods. Plos one, 14(6), e0218773. doi:10.1371/journal.pone.0218773

Pereira, R. H., Braga, C. K. V., Servo, L. M., Serra, B., Amaral, P., Gouveia, N., & Paez, A. (2021). Geographic access to COVID-19 healthcare in Brazil using a balanced float catchment area approach. Social Science & Medicine, 273. doi:10.1016/j.socscimed.2021.113773

Examples

library(accessibility)

# load a travel time matrix data in long format
data_path <- system.file("extdata/ttm_bho.rds", package = "accessibility")
ttm <- readRDS(data_path)

# 2SFCA with a step decay function
df <- floating_catchment_area(
       data = ttm,
       fca_metric = '2SFCA',
       orig_col = 'from_id',
       dest_col = 'to_id',
       opportunity_col = 'jobs',
       population_col = 'population',
       decay_function = decay_binary(cutoff = 50),
       travel_cost_col = 'travel_time'
       )
head(df)
#>            from_id access_2sfca
#> 1: 89a88cdb57bffff    0.4357418
#> 2: 89a88cdb597ffff    0.3938630
#> 3: 89a88cdb5b3ffff    0.4589910
#> 4: 89a88cdb5cfffff    0.5469433
#> 5: 89a88cd909bffff    0.4358530
#> 6: 89a88cd90b7ffff    0.5271746

# BFCA with an exponential decay function
df <- floating_catchment_area(
       data = ttm,
       fca_metric = 'BFCA',
       orig_col = 'from_id',
       dest_col = 'to_id',
       opportunity_col = 'jobs',
       population_col = 'population',
       decay_function = decay_exponential(decay_value = 0.5),
       travel_cost_col = 'travel_time'
       )
head(df)
#>            from_id access_bfca
#> 1: 89a88cdb57bffff  0.11894791
#> 2: 89a88cdb597ffff  0.24251324
#> 3: 89a88cdb5b3ffff  0.07459233
#> 4: 89a88cdb5cfffff  0.10470062
#> 5: 89a88cd909bffff  0.07549538
#> 6: 89a88cd90b7ffff  0.82935888