General impedance functions

Provides a numer of impedance functions to be used inside accessibility functions.

impedance_fun(t_ij, decay_function, cutoff = NULL, decay_value = NULL)

Arguments

t_ij

A number indicating the travel cost (time or money) between origin (i) and destination (j).

decay_function

A string. Which decay function to use when calculating accessibility. One of step, exponential, fixed_exponential, linear or logistic. Please see the details to understand how each alternative works and how they relate to the cutoffs and decay_value parameters.

cutoff

A numeric vector. This parameter has different effects for each decay function: it indicates the cutoff times in minutes when calculating cumulative opportunities accessibility with the step function...

decay_value

A number. Extra parameter to be passed to the selected decay_function. Has no effects when decay_function is either step or exponential.

Value

A function

Examples

library(accessibility)

# load a travel time matrix data in long format
data_path <- system.file("extdata/ttm_poa.csv", package = "accessibility")
ttm <- read.csv(data_path)

# Active accessibility: number of schools accessible from each origin
df <- cumulative_time_cutoff(data = ttm,
                            opportunity_colname = 'schools',
                            cutoff = 30,
                            by_colname = 'from_id')
head(df)
#>            from_id access
#> 1: 89a901291abffff      1
#> 2: 89a9012a3cfffff      0
#> 3: 89a901295b7ffff      2
#> 4: 89a901284a3ffff      0
#> 5: 89a9012809bffff      1
#> 6: 89a901285cfffff      3

# Passive accessibility: number of people that can reach each destination
df <- cumulative_time_cutoff(data = ttm,
                            opportunity_colname = 'population',
                            cutoff = 30,
                            by_colname = 'to_id')
head(df)
#>              to_id access
#> 1: 89a901291abffff   6082
#> 2: 89a90129157ffff   4920
#> 3: 89a9012956bffff   3657
#> 4: 89a90128257ffff  18011
#> 5: 89a90128043ffff  16033
#> 6: 89a90129ed3ffff   7434