Minimum travel time to closest opportunities

The function calculates the minimum travel time to closest N number of opportunities.

time_to_closest(data, opportunity_colname, by_colname, n_opportunities = 1)

Arguments

data

A data.frame with a travel time matrix in long format, containing the at least the columns of origin from_id, destination to_id, travel time travel_time from origin to destination, and number of opportunities in destination locations.

opportunity_colname

A string indicating the name of the column with data on the opportunities to be considered.

by_colname

A string with the name of the column of origin or destination that should be considered, indicating whether accessibility levels should by calculated by each origin (active accessibility) or destination (passive accessibility).

n_opportunities

A numeric value with the mininum N number of opportunities that should be considered. Defaults to 1

Value

A data.table object indicating for each origin the travel time to the closest opportunity and the id of the destination where it is located.

Examples

library(accessibility)

# load a travel time matrix data in long format
data_path <- system.file("extdata/ttm_bho.rds", package = "accessibility")
ttm <- readRDS(data_path)

df <- time_to_closest(data = ttm,
                     opportunity_colname = 'schools',
                     by_colname = 'from_id',
                     n_opportunities = 1)
head(df)
#>            from_id travel_time     destination
#> 1: 89a88cdb57bffff        28.0 89a88cdb53bffff
#> 2: 89a88cdb597ffff         5.8 89a88cdb597ffff
#> 3: 89a88cdb5b3ffff        11.0 89a88cdb5bbffff
#> 4: 89a88cdb5cfffff        13.0 89a88cdb5c7ffff
#> 5: 89a88cd909bffff         7.0 89a88cd9093ffff
#> 6: 89a88cd90b7ffff         6.0 89a88cd90a7ffff

df <- time_to_closest(data = ttm,
                     opportunity_colname = 'schools',
                     by_colname = 'from_id',
                     n_opportunities = 2)
head(df)
#>            from_id travel_time                     destination
#> 1: 89a881a5a2bffff          32 89a88cdb473ffff;89a88cdb0d3ffff
#> 2: 89a881a5a2fffff          NA                 89a88cdb593ffff
#> 3: 89a881a5a67ffff          31 89a88cdb473ffff;89a88cdb0d3ffff
#> 4: 89a881a5a6bffff          33                 89a88cdb5c7ffff
#> 5: 89a881a5a6fffff          35 89a88cdb473ffff;89a88cdb0d3ffff
#> 6: 89a881a5b03ffff          17 89a881a5b23ffff;89a881a5b37ffff