accessibility: Transport accessibility measures

A set of fast and convenient functions to help conducting accessibility analyses. Given a pre-computed travel cost matrix and a land use dataset (containing for example the location of jobs, healthcare and population), the package allows one to calculate accessibility levels, and accessibility poverty and inequality. The package covers the majority of the most commonly used accessibility measures (such as cumulative opportunities, gravity-based and floating catchment areas methods), some cutting edge measures proposed in the literature (e.g. balancing cost and constrained accessibility) as well as the most frequently used inequality and poverty metrics (such as the Palma ratio, the concentration and Theil indices and the FGT family of measures).

Usage

Please check the vignettes for more on the package usage:

Introduction to accessibility: calculating accessibility measures. Run vignette("accessibility") or check it on the website.

Decay functions. Run vignette("decay_functions", package = "accessibility") or check it on the website.

Calculating accessibility inequality and poverty. Run vignette("inequality_and_poverty", package = "accessibility") or check it on the website.

See also

Author

Maintainer: Daniel Herszenhut dhersz@gmail.com (ORCID)

Authors:

Rafael H. M. Pereira rafa.pereira.br@gmail.com (ORCID)

Anastasia Soukhov

Other contributors:

Joey Reid [contributor]

Ipea - Institute for Applied Economic Research [copyright holder, funder]