Cumulative access based on maximum travel time cutoff

The function calculates the number of opportunities accessible under a given travel time cutoff specified by the user.

cumulative_time_cutoff(
  data,
  cutoff,
  opportunity_col,
  travel_cost_col = "travel_time",
  by_col
)

Arguments

data

A data.frame with a travel cost matrix in long format, containing the columns with origin ids, destination ids, the travel cost between each origin-destination pair (this can be either in terms of travel time, distance or monetary costs), and number of population and opportunities located in each origin and destination.

cutoff

A numeric value indicating the maximum travel time considered.

opportunity_col

A string indicating the name of the column with the number of opportunities / resources / services to be considered.

travel_cost_col

A string with the name of the column with travel cost between origins and destinations. It can be either either travel time, distance or monetary costs. Defaults to travel_ime.

by_col

A string with the name of the column of origin or destination that should be considered, indicating whether accessibility levels should by calculated by each origin (active accessibility) or destination (passive accessibility).

Value

A data.table object.

See also

Examples

library(accessibility)

# load a travel time matrix data in long format
data_path <- system.file("extdata/ttm_bho.rds", package = "accessibility")
ttm <- readRDS(data_path)

# Active accessibility: number of schools accessible from each origin
df <- cumulative_time_cutoff(data = ttm,
                            cutoff = 30,
                            opportunity_col = 'schools',
                            travel_cost_col = 'travel_time',
                            by_col = 'from_id')
head(df)
#>            from_id access
#> 1: 89a88cdb57bffff      6
#> 2: 89a88cdb597ffff     22
#> 3: 89a88cdb5b3ffff     23
#> 4: 89a88cdb5cfffff      6
#> 5: 89a88cd909bffff     17
#> 6: 89a88cd90b7ffff     20

# Passive accessibility: number of people that can reach each destination
df <- cumulative_time_cutoff(data = ttm,
                            cutoff = 30,
                            opportunity_col = 'population',
                            travel_cost_col = 'travel_time',
                            by_col = 'to_id')
head(df)
#>              to_id access
#> 1: 89a88cdb57bffff  62591
#> 2: 89a88cdb19bffff 180316
#> 3: 89a88cdb1d3ffff 185293
#> 4: 89a88cdb56fffff 162202
#> 5: 89a88cdb0afffff 243124
#> 6: 89a88cdb50fffff 107605