Cumulative access based on time interval

The function calculates the average or median number of opportunities that can be reached considering multiple minute-by-minute time thresholds within a given travel time interval specified by the user.

cumulative_time_interval(
  data,
  opportunity_colname,
  start,
  end,
  by_colname,
  stat = "mean"
)

Arguments

data

A data.frame with a travel time matrix in long format, containing the at least the columns of origin from_id, destination to_id, travel time travel_time from origin to destination, and number of opportunities in destination locations.

opportunity_colname

A string indicating the name of the column with data on the opportunities to be considered.

start

An integer indicating the start point of the travel time interval.

end

An integer indicating the end point of the travel time interval.

by_colname

A string with the name of the column of origin or destination that should be considered, indicating whether accessibility levels should by calculated by each origin (active accessibility) or destination (passive accessibility).

stat

A string indicating the summary statistic used to aggregate the accessibility estimates within the time interval. It accepts either median (Default) or mean.

Value

A data.table object.

See also

Examples

library(accessibility)

# load a travel time matrix data in long format
data_path <- system.file("extdata/ttm_poa.csv", package = "accessibility")
ttm <- read.csv(data_path)

df <- cumulative_time_interval(data = ttm,
                              opportunity_colname = 'schools',
                              start = 20,
                              end = 30,
                              by_colname = 'from_id',
                              stat ='median')
head(df)
#>            from_id access
#> 1: 89a901291abffff      0
#> 2: 89a9012a3cfffff      0
#> 3: 89a901295b7ffff      1
#> 4: 89a901284a3ffff      0
#> 5: 89a9012809bffff      0
#> 6: 89a901285cfffff      3