Logistic-CDF decay function

Returns a logist-cdf weighting function parameterized with the median (inflection point) and standard deviation to be used inside accessibility calculating functions.

This function is generic over any kind of numeric travel cost, such as distance, time and money.

decay_logistic(median, sd)

Arguments

median

A numeric vector. The median or inflection point of the logistic CDF in minutes of travel time.

sd

A numeric vector with same length as median. The standard deviation in minutes of the logistic-CDF decay function must be greater than 0 and less than 120. Smaller values result in more rapid decay.

Value

A function that takes a generic travel cost vector (numeric) as an input and returns a vector of weights (numeric).

Details

When using a function created with decay_logistic(), the output is named after the combination of median ("m") and sd ("s")

e.g. given the median c(10, 20) and the sd c(10, 20), the outputs will be named "m(10);s(10)", "m(20);s(20)".

See also

Examples

weighting_function <- decay_logistic(
  median = seq(10, 120, by = 10),
  sd = 10
)
#> Error in decay_logistic(median = seq(10, 120, by = 10), sd = 10): unused argument (median = seq(10, 120, by = 10))

weighting_function(seq(0, 120, by = 5))
#> Error in weighting_function(seq(0, 120, by = 5)): could not find function "weighting_function"

weighting_function <- decay_logistic(
  c(10, 10, 10, 10, 20, 20, 20, 20),
  c(2, 4, 6, 8, 10, 12, 2, 4, 6, 8, 10, 12)
)
#> Error in decay_logistic(c(10, 10, 10, 10, 20, 20, 20, 20), c(2, 4, 6,     8, 10, 12, 2, 4, 6, 8, 10, 12)): Assertion on 'cutoff' failed: Contains duplicated values, position 2.

weighting_function(seq(0, 120, by = 5))
#> Error in weighting_function(seq(0, 120, by = 5)): could not find function "weighting_function"