Floating catchment area accessibilitt

Calculates accessibility accouting for the competition of resources using one of the multiple accessibility metrics in the floating catchment area family. The function currently includes 2SFCA and BFCA

floating_catchment_area(
  data,
  fca_metric,
  orig_col,
  dest_col,
  population_col,
  opportunity_col,
  decay_function
)

Arguments

data

A data.frame with a travel time matrix in long format, containing the at least the columns of origin from_id, destination to_id, travel time travel_time from origin to destination, and number of opportunities in destination locations.

fca_metric

A string indicating which floating cacthment area accessibility measure to use. Options avaiblable: "2SFCA" and "BFCA".

orig_col

A string with the name of the column of origin ids.

dest_col

A string with the name of the column of destination ids.

population_col

A string with the name of the column of origin with population count.

opportunity_col

A string indicating the name of the column with the number of opportunities / resources / services to be considered.

decay_function

A fuction that converts travel cost (t_ij) into an impedance factor used to weight opportunities. For convinence, the package currently includes the following functions: decay_bineary, decay_exponential, decay_exponential and decay_power.. See the documentation of each function for more details.

Value

A data.table object.

See also

Examples

library(accessibility)

# load a travel time matrix data in long format
data_path <- system.file("extdata/ttm_bho.rds", package = "accessibility")
ttm <- readRDS(data_path)

# 2SFCA with a step decay function
df <- floating_catchment_area(
       data = ttm,
       fca_metric = '2SFCA',
       orig_col = 'from_id',
       dest_col = 'to_id',
       opportunity_col = 'jobs',
       population_col = 'population',
       decay_function = decay_binary(cutoff = 50)
       )
head(df)
#>            from_id access_2sfca
#> 1: 89a88cdb57bffff    0.4357418
#> 2: 89a88cdb597ffff    0.3938630
#> 3: 89a88cdb5b3ffff    0.4589910
#> 4: 89a88cdb5cfffff    0.5469433
#> 5: 89a88cd909bffff    0.4358530
#> 6: 89a88cd90b7ffff    0.5271746

# BFCA with an exponential decay function
df <- floating_catchment_area(
       data = ttm,
       fca_metric = 'BFCA',
       orig_col = 'from_id',
       dest_col = 'to_id',
       opportunity_col = 'jobs',
       population_col = 'population',
       decay_function = decay_exponential(decay_value = 0.5)
       )
head(df)
#>            from_id access_bfca
#> 1: 89a88cdb57bffff  0.11894791
#> 2: 89a88cdb597ffff  0.24251324
#> 3: 89a88cdb5b3ffff  0.07459233
#> 4: 89a88cdb5cfffff  0.10470062
#> 5: 89a88cd909bffff  0.07549538
#> 6: 89a88cd90b7ffff  0.82935888