Floating catchment area accessibilitt

Calculates accessibility accouting for the competition of resources using one of the multiple accessibility metrics in the floating catchment area family. The function currently includes 2SFCA and BFCA

floating_catchment_area(
  data,
  fca_metric,
  orig_col,
  dest_col,
  population_col,
  opportunity_col,
  decay_function,
  cutoff = NULL,
  decay_value = NULL
)

Arguments

data

A data.frame with a travel time matrix in long format, containing the at least the columns of origin from_id, destination to_id, travel time travel_time from origin to destination, and number of opportunities in destination locations.

fca_metric

A string indicating which floating cacthment area accessibility measure to use. Options avaiblable: "2SFCA" and "BFCA".

orig_col

A string with the name of the column of origin ids.

dest_col

A string with the name of the column of destination ids.

population_col

A string with the name of the column of origin with population count.

opportunity_col

A string with the name of the column of destination with the number of opportunities / resources / services.

decay_function

A string. Which decay function to use when calculating accessibility. One of step, exponential, fixed_exponential, linear or logistic. Please see the details to understand how each alternative works and how they relate to the cutoffs and decay_value parameters.

cutoff

A numeric vector. This parameter has different effects for each decay function: it indicates the cutoff times in minutes when calculating cumulative opportunities accessibility with the step function...

decay_value

A number. Extra parameter to be passed to the selected decay_function. Has no effects when decay_function is either step or exponential.

Value

A numeric estimate of accessibility.

See also

Examples

library(accessibility)

# load a travel time matrix data in long format
data_path <- system.file("extdata/ttm_bho.rds", package = "accessibility")
ttm <- readRDS(data_path)

# 2SFCA with a step decay function
df <- floating_catchment_area(
       data = ttm,
       fca_metric = '2SFCA',
       orig_col = 'from_id',
       dest_col = 'to_id',
       opportunity_col = 'jobs',
       population_col = 'population',
       decay_function = 'step',
       cutoff = 30
       )
head(df)
#>            from_id access_2sfca
#> 1: 89a88cdb57bffff    0.1135253
#> 2: 89a88cdb597ffff    0.4412146
#> 3: 89a88cdb5b3ffff    0.4645848
#> 4: 89a88cdb5cfffff    0.3579043
#> 5: 89a88cd909bffff    0.2067682
#> 6: 89a88cd90b7ffff    0.2601425

# BFCA with a step decay function
df <- floating_catchment_area(
       data = ttm,
       fca_metric = 'BFCA',
       orig_col = 'from_id',
       dest_col = 'to_id',
       opportunity_col = 'jobs',
       population_col = 'population',
       decay_function = 'step',
       cutoff = 30
       )
head(df)
#>            from_id access_bfca
#> 1: 89a88cdb57bffff   0.1045731
#> 2: 89a88cdb597ffff   0.5069667
#> 3: 89a88cdb5b3ffff   0.5192616
#> 4: 89a88cdb5cfffff   0.4011974
#> 5: 89a88cd909bffff   0.2317942
#> 6: 89a88cd90b7ffff   0.2539631

df2 <- floating_catchment_area(data = ttm,
        fca_metric = '2SFCA',
        orig_col = 'from_id',
        dest_col = 'to_id',
        opportunity_col = 'jobs',
        population_col = 'population',
        decay_function = 'negative_exponential',
        decay_value = 0.5
        )

head(df2)
#>            from_id access_2sfca
#> 1: 89a88cdb57bffff   0.09719680
#> 2: 89a88cdb597ffff   0.25341503
#> 3: 89a88cdb5b3ffff   0.07677021
#> 4: 89a88cdb5cfffff   0.10922532
#> 5: 89a88cd909bffff   0.08838406
#> 6: 89a88cd90b7ffff   0.84150290