Minimum travel cost to closest N number of opportunities

The function calculates the minimum travel cost to closest N number of opportunities.

time_to_closest(
  data,
  opportunity_col,
  n_opportunities = 1,
  travel_cost_col = "travel_time",
  by_col
)

Arguments

data

A data.frame with a travel cost matrix in long format, containing the columns with origin ids, destination ids, the travel cost between each origin-destination pair (this can be either in terms of travel time, distance or monetary costs), and number of population and opportunities located in each origin and destination.

opportunity_col

A string indicating the name of the column with the number of opportunities / resources / services to be considered.

n_opportunities

A numeric value with the minimum N number of opportunities that should be considered. Defaults to 1

travel_cost_col

A string with the name of the column with travel cost between origins and destinations. It can be either either travel time, distance or monetary costs. Defaults to travel_ime.

by_col

A string with the name of the column of origin or destination that should be considered, indicating whether accessibility levels should by calculated by each origin (active accessibility) or destination (passive accessibility).

Value

A data.table object indicating for each origin the travel time to the closest opportunity and the id of the destination where it is located.

Examples

library(accessibility)

# load a travel time matrix data in long format
data_path <- system.file("extdata/ttm_bho.rds", package = "accessibility")
ttm <- readRDS(data_path)

df <- time_to_closest(data = ttm,
                     opportunity_col = 'schools',
                     n_opportunities = 1,
                     travel_cost_col = 'travel_time',
                     by_col = 'from_id')
head(df)
#>            from_id travel_cost     destination
#> 1: 89a88cdb57bffff        28.0 89a88cdb53bffff
#> 2: 89a88cdb597ffff         5.8 89a88cdb597ffff
#> 3: 89a88cdb5b3ffff        11.0 89a88cdb5bbffff
#> 4: 89a88cdb5cfffff        13.0 89a88cdb5c7ffff
#> 5: 89a88cd909bffff         7.0 89a88cd9093ffff
#> 6: 89a88cd90b7ffff         6.0 89a88cd90a7ffff

df <- time_to_closest(data = ttm,
                     opportunity_col = 'schools',
                     n_opportunities = 2,
                     travel_cost_col = 'travel_time',
                     by_col = 'from_id')
head(df)
#>            from_id travel_cost                     destination
#> 1: 89a881a5a2bffff          32 89a88cdb473ffff;89a88cdb0d3ffff
#> 2: 89a881a5a2fffff          NA                 89a88cdb593ffff
#> 3: 89a881a5a67ffff          31 89a88cdb473ffff;89a88cdb0d3ffff
#> 4: 89a881a5a6bffff          33                 89a88cdb5c7ffff
#> 5: 89a881a5a6fffff          35 89a88cdb473ffff;89a88cdb0d3ffff
#> 6: 89a881a5b03ffff          17 89a881a5b23ffff;89a881a5b37ffff